Added a scene with a skybox and asteroids(breakable, ice).
Basic spawning for scene objects is added, but should be updated later.
Breakable asteroid: If player hits this asteroid it breaks into smaller pieces.
Ice asteroid: Player can go into this asteroid and its vision is blurred for a while.
A sample blackhole is also added that can be improved later.